pyproject.toml
[tool.poetry]
name = "profile-as-code"
version = "0.1.0"
description = ""
authors = ["Theo Windebank <theo@theowindebank.co.uk>"]

[tool.poetry.dependencies]
python = "^3.8"
"ruamel.yaml" = "^0.17.17"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
